% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateYostVars.R
\name{calculateYostVars}
\alias{calculateYostVars}
\title{Calculate Yost Index component variables}
\usage{
calculateYostVars(data, acs_vars, shrink_sub = FALSE)
}
\arguments{
\item{data}{The data frame from `clean_acs_names`.}

\item{acs_vars}{The list of ACS variable names from the attribute.}

\item{shrink_sub}{Logical. If `TRUE`, also compute MOE columns for each variable.}
}
\value{
A data frame with the new component variables.
}
\description{
This internal helper function calculates the 7 component variables
for the Yost Index from the raw ACS data.
}
