% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAcsNames.R
\name{cleanAcsNames}
\alias{cleanAcsNames}
\title{Clean ACS geographic names}
\usage{
cleanAcsNames(dframe, geo)
}
\arguments{
\item{dframe}{The raw data from `fetchYostAcs`.}

\item{geo}{The geography level.}
}
\value{
A data frame with cleaned name columns.
}
\description{
This internal helper function splits the 'NAME' column from tidycensus
into clean columns for block group, tract, county, and state.
}
