% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchYostAcs.R
\name{fetchYostAcs}
\alias{fetchYostAcs}
\title{Fetch raw ACS data for Yost Index variables}
\usage{
fetchYostAcs(geo, year, states, get_geometry)
}
\arguments{
\item{geo}{Geography level (e.g., "county", "tract").}

\item{year}{The ACS5 year.}

\item{states}{A vector of state abbreviations or 'all'.}

\item{get_geometry}{Boolean, whether to download geometry.}
}
\value{
A tidycensus data frame.
}
\description{
This is an internal helper function that queries the Census API for all
variables required to compute the Yost Index.
}
