% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReproduceYost-package.R
\docType{package}
\name{ReproduceYost-package}
\alias{ReproduceYost}
\alias{ReproduceYost-package}
\title{ReproduceYost: A Tidy Interface for Calculating the Yost Index of Socioeconomic Status}
\description{
Provides a simple, tidy interface for calculating the Yost Index, a composite measure of neighborhood-level socioeconomic status (SES), using data from the US Census Bureau's American Community Survey (ACS). Features include automatic data fetching, flexible geographies (county, tract, block group), and optional spatial imputation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HanLabCollaboration/ReproduceYost}
  \item Report bugs at \url{https://github.com/HanLabCollaboration/ReproduceYost/issues}
}

}
\author{
\strong{Maintainer}: Rafael Catoia Pulgrossi \email{rcatoiap@stanford.edu}

Authors:
\itemize{
  \item Tae Yoon Lee \email{harrytyl@stanford.edu}
}

}
\keyword{internal}
