% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReproducibleYostIndex-package.R
\docType{package}
\name{ReproducibleYostIndex-package}
\alias{ReproducibleYostIndex}
\alias{ReproducibleYostIndex-package}
\title{ReproducibleYostIndex: Reproducing the Yost Index of Neighborhood Socioeconomic Status}
\description{
Provides a simple, tidy interface for calculating the Yost Index, a composite measure of neighborhood-level socioeconomic status (SES), using data from the US Census Bureau's American Community Survey (ACS). Features include automatic data fetching, flexible geographies (county, tract, block group), empirical Bayes stabilization, and optional spatial imputation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HanLabCollaboration/ReproducibleYostIndex}
  \item Report bugs at \url{https://github.com/HanLabCollaboration/ReproducibleYostIndex/issues}
}

}
\author{
\strong{Maintainer}: Rafael Catoia Pulgrossi \email{rcatoiap@stanford.edu}

Authors:
\itemize{
  \item Tae Yoon Lee \email{harrytyl@stanford.edu}
}

}
\keyword{internal}
